refactor(tasks): move use-tasks-fetch composable inside tasks store file
